Change concurrency level to be an int.  It removes conversion noise from the code

+++ stl/config.go
@@ -5,9 +5,9 @@ import (
 	"runtime"
 )
 
-var concurrencyLevel = uint32(runtime.NumCPU())
+var concurrencyLevel = runtime.NumCPU()
 
 // Allow users of stl package to override default concurrency level
-func SetConcurrencyLevel(l uint32) {
-	concurrencyLevel = uint32(math.Max(1, float64(l)))
+func SetConcurrencyLevel(l int) {
+	concurrencyLevel = int(math.Max(1, float64(l)))
 }
